move firebase to its own config file
